reverted long nonce (wrong branch)

+++ veriblock-core/src/main/java/org/veriblock/core/contracts/FilteredBlock.java
@@ -24,7 +24,7 @@ public interface FilteredBlock {
 
     int getDifficulty();
 
-    long getNonce();
+    int getNonce();
 
     int getTotalRegularTransactions();
